change double value for gson number deserializer, and change exception for file uploading while local cache is disabled.

+++ core/src/main/java/cn/leancloud/gson/NumberDeserializerDoubleAsIntFix.java
@@ -18,7 +18,7 @@ public class NumberDeserializerDoubleAsIntFix implements JsonDeserializer<Number
     // and return any type you want
     // this solution will transform 3.0 float to long values
     if (Math.floor(num.doubleValue()) != Math.ceil(num.doubleValue())) {
-      return num;
+      return num.doubleValue();
     }
     double doubleValue = Math.ceil(num.doubleValue());
     if (doubleValue == num.intValue()) {
